+++ vod/mkv/mkv_format.h
@@ -1,0 +1,10 @@
+#ifndef __MKV_FORMAT_H__
+#define __MKV_FORMAT_H__
+
+// includes
+#include "../media_format.h"
+
+// globals
+extern media_format_t mkv_format;
+
+#endif //__MKV_FORMAT_H__